added upcasts in fread pda and cfg for visual studio.

+++ grail/lib/io/fread_cfg.cpp
@@ -32,11 +32,11 @@ namespace grail { namespace io { namespace cfg {
 
     /// symbol character
     bool is_symbol_codepoint(const char * const cp) throw() {
-        return *cp < 0 || isalnum(*cp) || '_' == *cp;
+        return *cp < 0 || isalnum(static_cast<int>(*cp)) || '_' == *cp;
     }
 
     bool is_numeric_codepoint(const char * const cp) throw() {
-        return isdigit(*cp);
+        return isdigit(static_cast<int>(*cp));
     }
 
     uint8_t next_state(uint8_t curr_state, token_type input) throw() {
